+++ CHANGELOG.md
@@ -4,7 +4,23 @@ CHANGELOG
 3.1.0
 --
 
-Added feature to turn on/off preview of translated texts. When turned on, texts that are translated will be encapsulated by [].
+Added feature to turn on/off preview of translated texts. When turned on, texts that are translated will be encapsulated by user function.
+To enable it run this in console:
+
+```
+    b.spyTr(function(t){return "["+t+"]"});
+```
+
+To disable it run this in console:
+```
+    b.spyTr(null);
+```
+
+To check if enabled run this in console:
+```
+    b.spyTr()!==undefined
+```
+
 
 3.0.0
 --